cloud_server_requirement.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.0
Flask-PyMongo==0.5.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
paho-mqtt==1.3.0
pymongo==3.5.1
requests==2.18.4
SQLAlchemy==1.1.14
urllib3==1.22
visitor==0.1.3
Werkzeug==0.12.2
WTForms==2.1
